moviehub/requirements_dev.txt
-r requirements_minimum.txt
pytest==9.0.0
pytest-django==4.11.1
ipython==9.7.0
rich==14.2.0
django-debug-toolbar==6.2.0